requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
frozenlist==1.3.0
h11==0.13.0
idna==3.3
msgpack==1.0.3
multidict==6.0.2
outcome==1.1.0
py-algorand-sdk==1.9.0
py-cord==2.0.0b4879+gb830ca44
pycparser==2.21
pycryptodomex==3.13.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pytz==2021.3
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
table2ascii==0.1.4
tinyman-py-sdk==0.0.5
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
yarl==1.7.2
